+++ src/main/java/io/github/mityavasilyev/springreactbarapp/tag/TagController.java
@@ -1,5 +1,7 @@
 package io.github.mityavasilyev.springreactbarapp.tag;
 
+import io.github.mityavasilyev.springreactbarapp.exceptions.ExceptionController;
+import io.github.mityavasilyev.springreactbarapp.exceptions.InvalidIdException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -8,7 +10,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping(path = "api/tags")
-public class TagController {
+public class TagController extends ExceptionController {
 
     private final TagService tagService;
 
@@ -17,18 +19,19 @@ public class TagController {
     }
 
     @GetMapping
-    public List<Tag> getAllTags() {
-        return tagService.getAll();
+    public ResponseEntity<List<Tag>> getAllTags() {
+        return ResponseEntity.ok(tagService.getAll());
     }
 
     @GetMapping("{tagId}")
-    public Tag getTagById(@PathVariable("tagId") Long id) {
-        return tagService.getById(id);
+    public ResponseEntity<Tag> getTagById(@PathVariable("tagId") Long id) throws InvalidIdException {
+        if (id <= 0) throw new InvalidIdException();
+        return ResponseEntity.ok(tagService.getById(id));
     }
 
     @GetMapping("/name/{name}")
-    public List<Tag> getAllTagsByName(@PathVariable("name") String name) {
-        return tagService.getAllByName(name);
+    public ResponseEntity<List<Tag>> getAllTagsByName(@PathVariable("name") String name) {
+        return ResponseEntity.ok(tagService.getAllByName(name));
     }
 
     @PostMapping
@@ -39,14 +42,16 @@ public class TagController {
     }
 
     @DeleteMapping(path = "{tagId}")
-    public void deleteTag(@PathVariable("tagId") Long id) {
+    public ResponseEntity<Void> deleteTag(@PathVariable("tagId") Long id) throws InvalidIdException {
+        if (id <= 0) throw new InvalidIdException();
         tagService.deleteById(id);
+        return ResponseEntity.ok().build();
     }
 
-    // TODO: 31.01.2022 Update mappings to use ResponseEntity
     @PatchMapping(path = "{tagId}")
     public ResponseEntity<Tag> updateTag(@PathVariable("tagId") Long id,
-                                         @RequestBody TagDTO tagDTO) {
+                                         @RequestBody TagDTO tagDTO) throws InvalidIdException {
+        if (id <= 0) throw new InvalidIdException();
         Tag tagPatch = tagDTO.parseTagWithName();
         Tag tag = tagService.getById(id);
         if (tag == null) return ResponseEntity.notFound().build();
